Fixed file fetch mode and aliases not exporting
The *-CAS* aliases were there but not exporting, so they didn't work -
modified the exports to include any aliases. Also, in Get-MCASFile, the
fetch mode wasn't working properly because of a misplaced } on line 1485
that needed moved to line 1553. I must have broke that when i
implemented Invoke-MCASRestMethod and not realized it.